Because I don't know what the sigma was before...
I'm generating some more blue noise, this time non-triangular, and documenting the sigma for both.

This also fixes a bug in the Mitchell's best candidate code, where it would seed the RNG the same for every image, resulting in just one possible texture. This is likely a bug I introduced a long time ago, when trying to introduce specific patterns in the initial points.