Get everything using indexed-color for GL3 compat.
This uses a program from https://github.com/tommyettinger/TransparencyProcessor to convert lots of files to indexed-mode in a simple batch for-loop. Specifically, it uses the new -g option to use a grayscale palette with no transparency. The grayscale or alpha mode, specifically, has major problems with GL30 in libGDX, but indexed mode does not have problems.